avoid crash (CID 2449)
svn path=/trunk/KDE/kdebase/workspace/; revision=574559

+++ kdm/backend/ctrl.c
@@ -841,7 +841,7 @@ processCtrl( const char *string, int len, int fd, struct display *d )
 					}
 					break;
 				case remoteLogin:
-					if (di->serverPid != -1 && nuke)
+					if (di->serverPid != -1 && nuke && d)
 						TerminateProcess( di->serverPid, d->termSignal );
 					break;
 				case reserve:
